+++ frontend/app/src/store/balances/types.d.ts
@@ -184,3 +184,5 @@ export interface ERC20Token {
   readonly name?: string;
   readonly symbol?: string;
 }
+
+export type ExchangeRateGetter = (currency: string) => BigNumber | undefined;