+++ frontend/app/tests/unit/store/settings/utils.spec.ts
@@ -1,3 +1,4 @@
+import { default as BigNumber } from 'bignumber.js';
 import { Defaults } from '@/data/defaults';
 import {
   TIMEFRAME_SETTING,
@@ -13,7 +14,9 @@ import {
   CURRENCY_LOCATION,
   REFRESH_PERIOD,
   EXPLORERS,
-  ITEMS_PER_PAGE
+  ITEMS_PER_PAGE,
+  AMOUNT_ROUNDING_MODE,
+  VALUE_ROUNDING_MODE
 } from '@/store/settings/consts';
 import { loadFrontendSettings } from '@/store/settings/utils';
 
@@ -66,7 +69,9 @@ describe('settings:utils', () => {
         [CURRENCY_LOCATION]: Defaults.DEFAULT_CURRENCY_LOCATION,
         [REFRESH_PERIOD]: -1,
         [EXPLORERS]: {},
-        [ITEMS_PER_PAGE]: 10
+        [ITEMS_PER_PAGE]: 10,
+        [AMOUNT_ROUNDING_MODE]: BigNumber.ROUND_UP,
+        [VALUE_ROUNDING_MODE]: BigNumber.ROUND_DOWN
       },
       { root: true }
     );